Empty Path For Recipes with No Image
Previously recipes with no image would fail to load,
with this change their URL will be an empty URL. But
more than that, since an object is returned the
resize calls against the image will no longer error
and an empty string will still be returned. The view
layer can handle the empty string and an acceptable
view.