Only show current-src when its present
Currently, the image urlBuilder creates a
path based route to the URL. When an image
url is provided to the image-upload component
we rely only a config value to prefix the
domain. This dependency was causing issues
with passing an empty string to the component
property, mainnly that it would at minimum
receive the the host. the currentSrc is now
only shown when their is an image associated
with the recipe.