Only Show Image Styles for Recipes With Lede Image
Previously the styles rendering a full screen background
where displayed even when the recipe did not have an image
this cause the recipe to look empty as only a gray screen
was displayed.

This also includes other cosmetic changes. The directions
label was changed from 'What To Do' as it seemed more serious.
mdash support was added for serving size as well as cook time.